+++ web/src/app/app.routes.ts
@@ -5,5 +5,5 @@ import {Register} from './features/register/register';
 export const routes: Routes = [
   {path: "login", component: Login},
   {path: "register", component: Register},
-  {path: "driver-history", loadComponent: () => import('./features/ride-history/driver-history.component').then(m => m.DriverHistory)},
+  {path: "driver-history", loadComponent: () => import('./features/ride-history/driver-history').then(m => m.DriverHistory)},
 ];
