+++ web/src/app/features/ride-history/driver-history.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { DriverHistory } from './driver-history.component';
+import { DriverHistory } from './driver-history';
 
 describe('DriverHistory', () => {
   let component: DriverHistory;
